pyproject.toml
[project]
name = "threat-intel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "google-genai>=1.59.0",
    "ipython>=9.9.0",
    "loguru>=0.7.3",
    "playwright>=1.57.0",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy[asyncio]>=2.0.45",
    "tenacity>=9.1.2",
]
